app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bidict==0.22.1
click==8.1.7
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
